requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
larksuite-oapi>=1.0.30
pandas>=2.2.0
openpyxl>=3.1.0
httpx>=0.27.0
python-dotenv>=1.0.0
